Move contact form and its deps into a new component
