+++ src/pages_styles/index.styles.ts
@@ -285,46 +285,6 @@ export const StyledProjects = styled.div`
 `;
 
 
-// CONTACT
-export const StyledContactSection = styled.form`
-  width: 90%;
-  margin: 0 auto;
-  display: flex;
-  flex-direction: column;
-  position: relative;
-
-  @media ${media.tablet}{
-    width: 80%;
-  }
-
-  @media ${media.laptop}{
-    width: 75%;
-  }
-`;
-
-export const StyledContactSectionButtonWrapper = styled.div`
-  margin: 2rem auto 0 auto;
-`;
-
-export const StyledContactThankWrapper = styled.div`
-  margin-top: -2rem;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-`;
-
-export const StyledContactThankIcon = styled.img`
-  display: block;
-  width: 10rem;
-  margin-bottom: 2rem;
-
-  @media ${media.laptop}{
-    width: 15rem;
-  }
-`;
-
-
 // MISC
 export const StyledScrollupButton = styled.button`
   display: none;
